Created function to read information from MongoDB on server startup
